Move LLMChoice to message module
